CLN: move core/formats.py and core/style.py to formats
