+++ pandas/formats/format.py
