+++ pandas/formats/style.py
