+++ pandas/indexes/base.py
@@ -609,7 +609,7 @@ class Index(IndexOpsMixin, StringAccessorMixin, PandasObject):
         """
         Return the formatted data as a unicode string
         """
-        from pandas.core.format import get_console_size, _get_adjustment
+        from pandas.formats.format import get_console_size, _get_adjustment
         display_width, _ = get_console_size()
         if display_width is None:
             display_width = get_option('display.width') or 80
@@ -1456,7 +1456,7 @@ class Index(IndexOpsMixin, StringAccessorMixin, PandasObject):
     def _format_with_header(self, header, na_rep='NaN', **kwargs):
         values = self.values
 
-        from pandas.core.format import format_array
+        from pandas.formats.format import format_array
 
         if is_categorical_dtype(values.dtype):
             values = np.array(values)
