+++ pandas/io/tests/test_excel.py
@@ -1826,7 +1826,7 @@ class Openpyxl20Tests(ExcelWriterBase, tm.TestCase):
         self.assertEqual(kw['protection'], protection)
 
     def test_write_cells_merge_styled(self):
-        from pandas.core.format import ExcelCell
+        from pandas.formats.format import ExcelCell
         from openpyxl import styles
 
         sheet_name = 'merge_styled'
@@ -1939,7 +1939,7 @@ class Openpyxl22Tests(ExcelWriterBase, tm.TestCase):
         if not openpyxl_compat.is_compat(major_ver=2):
             raise nose.SkipTest('incompatiable openpyxl version')
 
-        from pandas.core.format import ExcelCell
+        from pandas.formats.format import ExcelCell
 
         sheet_name = 'merge_styled'
